# services/matchmaking-service/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for Valorant teammate matchmaking. The service creates matchmaking requests, retrieves candidates from the Python Player Service via gRPC, scores and filters them, and creates match groups.",
        "title": "PushA Matchmaking Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/matchmaking/requests": {
            "post": {
                "description": "Creates a new Valorant matchmaking request for a player.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matchmaking Requests"
                ],
                "summary": "Create matchmaking request",
                "parameters": [
                    {
                        "description": "Create matchmaking request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMatchmakingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.MatchmakingRequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/matchmaking/requests/{request_id}": {
            "get": {
                "description": "Returns a matchmaking request by its identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matchmaking Requests"
                ],
                "summary": "Get matchmaking request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matchmaking request ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MatchmakingRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/matchmaking/requests/{request_id}/candidates": {
            "get": {
                "description": "Returns candidates previously found and stored for the specified matchmaking request.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Candidates"
                ],
                "summary": "Get candidates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matchmaking request ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchCandidatesResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/matchmaking/requests/{request_id}/group": {
            "post": {
                "description": "Creates a match group using the request author and selected candidate IDs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Match Groups"
                ],
                "summary": "Create match group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matchmaking request ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Selected candidate IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMatchGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.MatchGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/matchmaking/requests/{request_id}/search": {
            "post": {
                "description": "Retrieves players from the Python Player Service via gRPC, filters them, calculates scores and stores candidates.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Candidates"
                ],
                "summary": "Search candidates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matchmaking request ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchCandidatesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/players/{player_id}/matchmaking/requests": {
            "get": {
                "description": "Returns all matchmaking requests created by the specified player.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matchmaking Requests"
                ],
                "summary": "Get player matchmaking requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player ID",
                        "name": "player_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.MatchmakingRequestResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CandidateResponse": {
            "type": "object",
            "properties": {
                "current_rank": {
                    "type": "string"
                },
                "main_roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nickname": {
                    "type": "string"
                },
                "player_id": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "riot_id": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "teammate_rating": {
                    "type": "number"
                }
            }
        },
        "dto.CreateMatchGroupRequest": {
            "type": "object",
            "properties": {
                "selected_candidate_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.CreateMatchmakingRequest": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "max_rank": {
                    "type": "string"
                },
                "min_rank": {
                    "type": "string"
                },
                "min_teammate_rating": {
                    "type": "number"
                },
                "needed_players": {
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "required_player_status": {
                    "type": "string"
                },
                "required_roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "strategy": {
                    "type": "string"
                }
            }
        },
        "dto.MatchGroupResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.MatchmakingRequestResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_rank": {
                    "type": "string"
                },
                "min_rank": {
                    "type": "string"
                },
                "min_teammate_rating": {
                    "type": "number"
                },
                "needed_players": {
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "required_player_status": {
                    "type": "string"
                },
                "required_roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "strategy": {
                    "type": "string"
                }
            }
        },
        "dto.SearchCandidatesResponse": {
            "type": "object",
            "properties": {
                "candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CandidateResponse"
                    }
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                }
            }
        }
    }
}